feat: Add course on sustainability and technology
This commit adds a new course titled "Como Impulsionar a Sustentabilidade Através da Tecnologia" to the list of courses. The course is from the institution "Linkedin Learning" and can be accessed through the certificate URL provided. It is tagged with "sustainability" and "technology" and has a conclusion date of September 16, 2024.